part4/requirements.txt
flask
flask-restx
Flask-SQLAlchemy
flask-bcrypt
flask-jwt-extended
flask-cors